refactor: HttpRequestHandler 라우트테이블 추가
